+++ docs/Whats-New.md
@@ -372,6 +372,7 @@ New:
 - Airstrike & spy plane fixed spawn distance & height (by Starkku)
 - Web logic against infantry (FS-21)
 - Remove parasites in any warhead (by FS-21)
+- Script action for repairing destroyed bridges (by FS-21)
 - Allow enabling application of `Verses` and `PercentAtMax` for negative damage (by Starkku)
 - `Burst.Retarget` for assigning a new target in each projectile (by FS-21)
 - Script actions for manipulating script flows with conditional jumps (by FS-21)